Added initial odin output interface: tripdirections.proto Added skeleton code for builders Added total distance for simple output
