+++ src/odin/maneuversbuilder.cc
@@ -1,0 +1,16 @@
+#include <iostream>
+
+#include "odin/maneuversbuilder.h"
+
+namespace valhalla {
+namespace odin {
+
+ManeuversBuilder::ManeuversBuilder(TripPath& trip_path)
+    : trip_path_(trip_path) {
+}
+
+void ManeuversBuilder::Build() {
+}
+
+}
+}
